Move mesh and primitives to submodules, clean up helpers
